+++ cl/compile_test.go
@@ -54,292 +54,3 @@ func TestBasic(t *testing.T) {
 }
 
 // -----------------------------------------------------------------------------
-
-var fsTestUnbound = asttest.NewSingleFileFS("/foo", "bar.ql", `
-	println("Hello " + "qiniu:", 123, 4.5, 7i)
-`)
-
-func TestUnbound(t *testing.T) {
-	fset := token.NewFileSet()
-	pkgs, err := parser.ParseFSDir(fset, fsTestUnbound, "/foo", nil, 0)
-	if err != nil || len(pkgs) != 1 {
-		t.Fatal("ParseFSDir failed:", err, len(pkgs))
-	}
-
-	bar := pkgs["main"]
-	b := exec.NewBuilder(nil)
-	_, err = NewPackage(b, bar)
-	if err != nil {
-		t.Fatal("Compile failed:", err)
-	}
-	code := b.Resolve()
-
-	ctx := exec.NewContext(code)
-	ctx.Exec(0, code.Len())
-	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
-	if v := ctx.Get(-1); v != nil {
-		t.Fatal("error:", v)
-	}
-	if v := ctx.Get(-2); v != int(28) {
-		t.Fatal("n:", v)
-	}
-}
-
-// -----------------------------------------------------------------------------
-
-var fsTestOperator = asttest.NewSingleFileFS("/foo", "bar.ql", `
-	println("Hello", 123 * 4.5, 1 + 7i)
-`)
-
-func TestOperator(t *testing.T) {
-	fset := token.NewFileSet()
-	pkgs, err := parser.ParseFSDir(fset, fsTestOperator, "/foo", nil, 0)
-	if err != nil || len(pkgs) != 1 {
-		t.Fatal("ParseFSDir failed:", err, len(pkgs))
-	}
-
-	bar := pkgs["main"]
-	b := exec.NewBuilder(nil)
-	_, err = NewPackage(b, bar)
-	if err != nil {
-		t.Fatal("Compile failed:", err)
-	}
-	code := b.Resolve()
-
-	ctx := exec.NewContext(code)
-	ctx.Exec(0, code.Len())
-	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
-	if v := ctx.Get(-1); v != nil {
-		t.Fatal("error:", v)
-	}
-	if v := ctx.Get(-2); v != int(19) {
-		t.Fatal("n:", v)
-	}
-}
-
-// -----------------------------------------------------------------------------
-
-var fsTestVar = asttest.NewSingleFileFS("/foo", "bar.ql", `
-	x := 123.1
-	println("Hello", x)
-`)
-
-func TestVar(t *testing.T) {
-	fset := token.NewFileSet()
-	pkgs, err := parser.ParseFSDir(fset, fsTestVar, "/foo", nil, 0)
-	if err != nil || len(pkgs) != 1 {
-		t.Fatal("ParseFSDir failed:", err, len(pkgs))
-	}
-
-	bar := pkgs["main"]
-	b := exec.NewBuilder(nil)
-	_, err = NewPackage(b, bar)
-	if err != nil {
-		t.Fatal("Compile failed:", err)
-	}
-	code := b.Resolve()
-
-	ctx := exec.NewContext(code)
-	ctx.Exec(0, code.Len())
-	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
-	if v := ctx.Get(-1); v != nil {
-		t.Fatal("error:", v)
-	}
-	if v := ctx.Get(-2); v != int(12) {
-		t.Fatal("n:", v)
-	}
-}
-
-// -----------------------------------------------------------------------------
-
-var fsTestVarOp = asttest.NewSingleFileFS("/foo", "bar.ql", `
-	x := 123.1
-	y := 1 + x
-	println("Hello", y + 10)
-	n, err := println("Hello", y + 10)
-	println("ret:", n << 1, err)
-`)
-
-func TestVarOp(t *testing.T) {
-	fset := token.NewFileSet()
-	pkgs, err := parser.ParseFSDir(fset, fsTestVarOp, "/foo", nil, 0)
-	if err != nil || len(pkgs) != 1 {
-		t.Fatal("ParseFSDir failed:", err, len(pkgs))
-	}
-
-	bar := pkgs["main"]
-	b := exec.NewBuilder(nil)
-	_, err = NewPackage(b, bar)
-	if err != nil {
-		t.Fatal("Compile failed:", err)
-	}
-	code := b.Resolve()
-
-	ctx := exec.NewContext(code)
-	ctx.Exec(0, code.Len())
-	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
-	if v := ctx.Get(-1); v != nil {
-		t.Fatal("error:", v)
-	}
-	if v := ctx.Get(-2); v != int(14) {
-		t.Fatal("n:", v)
-	}
-}
-
-// -----------------------------------------------------------------------------
-
-var fsTestGoPackage = asttest.NewSingleFileFS("/foo", "bar.ql", `
-	import "fmt"
-	import gostrings "strings"
-
-	x := gostrings.NewReplacer("?", "!").Replace("hello, world???")
-	fmt.Println("x: " + x)
-`)
-
-func TestGoPackage(t *testing.T) {
-	fset := token.NewFileSet()
-	pkgs, err := parser.ParseFSDir(fset, fsTestGoPackage, "/foo", nil, 0)
-	if err != nil || len(pkgs) != 1 {
-		t.Fatal("ParseFSDir failed:", err, len(pkgs))
-	}
-
-	bar := pkgs["main"]
-	b := exec.NewBuilder(nil)
-	_, err = NewPackage(b, bar)
-	if err != nil {
-		t.Fatal("Compile failed:", err)
-	}
-	code := b.Resolve()
-
-	ctx := exec.NewContext(code)
-	ctx.Exec(0, code.Len())
-	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
-	if v := ctx.Get(-1); v != nil {
-		t.Fatal("error:", v)
-	}
-	if v := ctx.Get(-2); v != int(19) {
-		t.Fatal("n:", v)
-	}
-}
-
-// -----------------------------------------------------------------------------
-
-var fsTestReturn = asttest.NewSingleFileFS("/foo", "bar.ql", `
-	import (
-		"fmt"
-		"strings"
-	)
-
-	func foo(x string) string {
-		return strings.NewReplacer("?", "!").Replace(x)
-	}
-
-	fmt.Println(foo("Hello, world???"))
-`)
-
-func TestReturn(t *testing.T) {
-	fset := token.NewFileSet()
-	pkgs, err := parser.ParseFSDir(fset, fsTestReturn, "/foo", nil, 0)
-	if err != nil || len(pkgs) != 1 {
-		t.Fatal("ParseFSDir failed:", err, len(pkgs))
-	}
-
-	bar := pkgs["main"]
-	b := exec.NewBuilder(nil)
-	_, err = NewPackage(b, bar)
-	if err != nil {
-		t.Fatal("Compile failed:", err)
-	}
-	code := b.Resolve()
-
-	ctx := exec.NewContext(code)
-	ctx.Exec(0, code.Len())
-	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
-	if v := ctx.Get(-1); v != nil {
-		t.Fatal("error:", v)
-	}
-	if v := ctx.Get(-2); v != int(16) {
-		t.Fatal("n:", v)
-	}
-}
-
-// -----------------------------------------------------------------------------
-
-var fsTestFunc = asttest.NewSingleFileFS("/foo", "bar.ql", `
-	import "fmt"
-
-	func foo(x string) (n int, err error) {
-		n, err = fmt.Println("x: " + x)
-		return
-	}
-
-	foo("Hello, world!")
-`)
-
-func TestFunc(t *testing.T) {
-	fset := token.NewFileSet()
-	pkgs, err := parser.ParseFSDir(fset, fsTestFunc, "/foo", nil, 0)
-	if err != nil || len(pkgs) != 1 {
-		t.Fatal("ParseFSDir failed:", err, len(pkgs))
-	}
-
-	bar := pkgs["main"]
-	b := exec.NewBuilder(nil)
-	_, err = NewPackage(b, bar)
-	if err != nil {
-		t.Fatal("Compile failed:", err)
-	}
-	code := b.Resolve()
-
-	ctx := exec.NewContext(code)
-	ctx.Exec(0, code.Len())
-	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
-	if v := ctx.Get(-1); v != nil {
-		t.Fatal("error:", v)
-	}
-	if v := ctx.Get(-2); v != int(17) {
-		t.Fatal("n:", v)
-	}
-}
-
-// -----------------------------------------------------------------------------
-
-var fsTestClosure = asttest.NewSingleFileFS("/foo", "bar.ql", `
-	import "fmt"
-
-	foo := func(prompt string) (n int, err error) {
-		n, err = fmt.Println(prompt + x)
-		return
-	}
-
-	x := "Hello, world!"
-	foo("x: ")
-`)
-
-func TestClosure(t *testing.T) {
-	fset := token.NewFileSet()
-	pkgs, err := parser.ParseFSDir(fset, fsTestClosure, "/foo", nil, 0)
-	if err != nil || len(pkgs) != 1 {
-		t.Fatal("ParseFSDir failed:", err, len(pkgs))
-	}
-
-	bar := pkgs["main"]
-	b := exec.NewBuilder(nil)
-	_, err = NewPackage(b, bar)
-	if err != nil {
-		t.Fatal("Compile failed:", err)
-	}
-	code := b.Resolve()
-
-	ctx := exec.NewContext(code)
-	ctx.Exec(0, code.Len())
-	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
-	if v := ctx.Get(-1); v != nil {
-		t.Fatal("error:", v)
-	}
-	if v := ctx.Get(-2); v != int(17) {
-		t.Fatal("n:", v)
-	}
-}
-
-// -----------------------------------------------------------------------------
