+++ cl/stmt_test.go
@@ -1,0 +1,133 @@
+package cl
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/qiniu/qlang/ast/asttest"
+	"github.com/qiniu/qlang/exec"
+	"github.com/qiniu/qlang/parser"
+	"github.com/qiniu/qlang/token"
+)
+
+// -----------------------------------------------------------------------------
+
+var fsTestReturn = asttest.NewSingleFileFS("/foo", "bar.ql", `
+	import (
+		"fmt"
+		"strings"
+	)
+
+	func foo(x string) string {
+		return strings.NewReplacer("?", "!").Replace(x)
+	}
+
+	fmt.Println(foo("Hello, world???"))
+`)
+
+func TestReturn(t *testing.T) {
+	fset := token.NewFileSet()
+	pkgs, err := parser.ParseFSDir(fset, fsTestReturn, "/foo", nil, 0)
+	if err != nil || len(pkgs) != 1 {
+		t.Fatal("ParseFSDir failed:", err, len(pkgs))
+	}
+
+	bar := pkgs["main"]
+	b := exec.NewBuilder(nil)
+	_, err = NewPackage(b, bar)
+	if err != nil {
+		t.Fatal("Compile failed:", err)
+	}
+	code := b.Resolve()
+
+	ctx := exec.NewContext(code)
+	ctx.Exec(0, code.Len())
+	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
+	if v := ctx.Get(-1); v != nil {
+		t.Fatal("error:", v)
+	}
+	if v := ctx.Get(-2); v != int(16) {
+		t.Fatal("n:", v)
+	}
+}
+
+// -----------------------------------------------------------------------------
+
+var fsTestFunc = asttest.NewSingleFileFS("/foo", "bar.ql", `
+	import "fmt"
+
+	func foo(x string) (n int, err error) {
+		n, err = fmt.Println("x: " + x)
+		return
+	}
+
+	foo("Hello, world!")
+`)
+
+func TestFunc(t *testing.T) {
+	fset := token.NewFileSet()
+	pkgs, err := parser.ParseFSDir(fset, fsTestFunc, "/foo", nil, 0)
+	if err != nil || len(pkgs) != 1 {
+		t.Fatal("ParseFSDir failed:", err, len(pkgs))
+	}
+
+	bar := pkgs["main"]
+	b := exec.NewBuilder(nil)
+	_, err = NewPackage(b, bar)
+	if err != nil {
+		t.Fatal("Compile failed:", err)
+	}
+	code := b.Resolve()
+
+	ctx := exec.NewContext(code)
+	ctx.Exec(0, code.Len())
+	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
+	if v := ctx.Get(-1); v != nil {
+		t.Fatal("error:", v)
+	}
+	if v := ctx.Get(-2); v != int(17) {
+		t.Fatal("n:", v)
+	}
+}
+
+// -----------------------------------------------------------------------------
+
+var fsTestClosure = asttest.NewSingleFileFS("/foo", "bar.ql", `
+	import "fmt"
+
+	foo := func(prompt string) (n int, err error) {
+		n, err = fmt.Println(prompt + x)
+		return
+	}
+
+	x := "Hello, world!"
+	foo("x: ")
+`)
+
+func TestClosure(t *testing.T) {
+	fset := token.NewFileSet()
+	pkgs, err := parser.ParseFSDir(fset, fsTestClosure, "/foo", nil, 0)
+	if err != nil || len(pkgs) != 1 {
+		t.Fatal("ParseFSDir failed:", err, len(pkgs))
+	}
+
+	bar := pkgs["main"]
+	b := exec.NewBuilder(nil)
+	_, err = NewPackage(b, bar)
+	if err != nil {
+		t.Fatal("Compile failed:", err)
+	}
+	code := b.Resolve()
+
+	ctx := exec.NewContext(code)
+	ctx.Exec(0, code.Len())
+	fmt.Println("results:", ctx.Get(-2), ctx.Get(-1))
+	if v := ctx.Get(-1); v != nil {
+		t.Fatal("error:", v)
+	}
+	if v := ctx.Get(-2); v != int(17) {
+		t.Fatal("n:", v)
+	}
+}
+
+// -----------------------------------------------------------------------------
